backend/tagger/pyproject.toml
[project]
name = "tagger"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "alex-database",
    "alex-llm",
    "boto3>=1.40.9",
    "langfuse>=3.3.4",
    "openai-agents[litellm]>=0.3.0",
    "pydantic>=2.11.7",
    "pydantic-ai>=1.0.6",
    "python-dotenv>=1.1.1",
    "tenacity>=9.1.2",
]

[tool.uv.sources]
alex-database = { path = "../database", editable = true }
alex-llm = { path = "../alex_llm", editable = true }
